Fix: cancel releases in progress when a new one comes in (#48)
There is actually no need to wait for a release to finish before
starting the new one. Worst case, the deployment was busy, but
Nomad can handle it just fine if two deployments arrive one after
the other.

But more likely it cancels the building of the Dockers, which is
totally fine. There, worst case, we pushed some dangling images
to the container registry.